requirements.txt
beautifulsoup4==4.9.0
boto3
channels==3.0.3
codeguru_profiler_agent
color-it==2.1.3
cryptocode
Deprecated==1.2.10
Django==3.0.4
django-cors-headers==3.7.0
django-middleware-global-request==0.2.0
django-extensions==2.2.9
django-prometheus
django-storages
durations==0.3.3
Jinja2
Pillow==7.2.0
prometheus-client
PyJWT==2.0.1
PyYAML==3.13
pytest-django==4.1.0
pytest-mock==3.4.0
python-gitlab==2.6.0
python-keycloak==0.23.0
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.4.5
textdistance==4.2.0
wheel==0.34.2
xmljson==0.2.1
xmltodict==0.12.0

gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
# psycopg2==2.8.4 # uncomment if you prefer to build from source
psycopg2-binary==2.8.6
uvicorn
